requirements.txt
# FastAPI and ASGI server
fastapi[standard]
uvicorn

# Google authentication (OAuth2 verification)
google-auth
google-auth-oauthlib
google-auth-httplib2

# HTTP requests
requests

# Pydantic (data validation)
pydantic

# MySQL 
mysql-connector-python
PyMySQL
cryptography

# SQLAlchemy
sqlalchemy

# Alembic (para migraciones)
alembic

# leer de entorno
python-dotenv
